Added all the Wwise plugins to my project.
Sounds are working.

Disabled playing with audio. It was throwing errors.

Noticed that the timer does not reset after the game is
over. Resulting in an infinite game over loop.

Game Needs redesign of some of the layouts. Some of the obstacles dont
make sense. Also the player can get stuck in some of them.